// extensions/forscore/src/types/Score.ts
export interface Score {
  filename: string;
  title: string;
  startPage?: string;
  endPage?: string;
  composers?: string;
  genres?: string;
  tags?: string;
  labels?: string;
  id?: string;
  rating?: string;
  difficulty?: string;
  minutes?: string;
  seconds?: string;
  keysf?: string;
  keymi?: string;
}